june-2021-leetcoding-challenge/src/test/java/org/redquark/leetcode/challenge/Problem01_MaxAreaOfIslandTest.java
package org.redquark.leetcode.challenge;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Problem01_MaxAreaOfIslandTest {

    private final Problem01_MaxAreaOfIsland testObject = new Problem01_MaxAreaOfIsland();

    @Test
    void testMaxAreaOfIsland() {
        int[][] grid = new int[][]{
                {0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0},
                {0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0},
                {0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0},
                {0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0},
                {0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0}
        };
        assertEquals(6, testObject.maxAreaOfIsland(grid));

        grid = new int[][]{
                {0, 0, 0, 0, 0, 0, 0, 0}
        };
        assertEquals(0, testObject.maxAreaOfIsland(grid));

    }
}